# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * AnonMatch API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface User {
  id: number;
  clerkId: string;
  displayName: string;
  /** @nullable */
  username?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  bio?: string | null;
  isVip: boolean;
  isOnline: boolean;
  isSearching: boolean;
  createdAt: string;
}

export interface UserUpdate {
  displayName?: string;
  username?: string;
  bio?: string;
}

export interface UserStats {
  totalChats: number;
  totalMessages: number;
  totalReveals: number;
  postsCount: number;
}

export interface AvatarUpload {
  file: string;
}

export interface AvatarUploadResult {
  avatarUrl: string;
}

export interface Post {
  id: number;
  authorId: number;
  authorDisplayName: string;
  /** @nullable */
  authorAvatarUrl?: string | null;
  content: string;
  /** @nullable */
  imageUrl?: string | null;
  isBlurred: boolean;
  isOwn?: boolean;
  createdAt: string;
}

export interface PostList {
  posts: Post[];
  hasMore: boolean;
  /** @nullable */
  nextCursor?: number | null;
}

export interface PostInput {
  /** @minLength 1 */
  content: string;
}

export interface ImageUpload {
  file: string;
}

export interface ImageUploadResult {
  imageUrl: string;
}

export type SearchResultStatus = typeof SearchResultStatus[keyof typeof SearchResultStatus];


export const SearchResultStatus = {
  searching: 'searching',
  matched: 'matched',
} as const;

export type MatchStateStatus = typeof MatchStateStatus[keyof typeof MatchStateStatus];


export const MatchStateStatus = {
  active: 'active',
  ended: 'ended',
} as const;

export interface MatchState {
  id: number;
  partnerId: number;
  partnerDisplayName: string;
  /** @nullable */
  partnerAvatarUrl?: string | null;
  myMessageCount: number;
  partnerMessageCount: number;
  myRevealRequested: boolean;
  partnerRevealRequested: boolean;
  isRevealed: boolean;
  status: MatchStateStatus;
  createdAt: string;
  /** @nullable */
  endedAt?: string | null;
}

export interface SearchResult {
  status: SearchResultStatus;
  match?: MatchState;
}

export interface Message {
  id: number;
  matchId: number;
  senderId: number;
  content: string;
  isMine: boolean;
  createdAt: string;
}

export interface MessageList {
  messages: Message[];
  hasMore: boolean;
  /** @nullable */
  nextCursor?: number | null;
}

export interface MessageInput {
  /** @minLength 1 */
  content: string;
}

export interface RevealResult {
  myRevealRequested: boolean;
  partnerRevealRequested: boolean;
  isRevealed: boolean;
}

export interface MatchHistoryItem {
  id: number;
  partnerDisplayName: string;
  /** @nullable */
  partnerAvatarUrl?: string | null;
  myMessageCount: number;
  partnerMessageCount: number;
  isRevealed: boolean;
  createdAt: string;
  /** @nullable */
  endedAt?: string | null;
}

export interface MatchHistoryList {
  matches: MatchHistoryItem[];
}

export type ListPostsParams = {
/**
 * Cursor for pagination (last post id)
 * @nullable
 */
cursor?: number | null;
/**
 * Number of posts to return (default 20)
 * @nullable
 */
limit?: number | null;
};

export type ListMessagesParams = {
/**
 * @nullable
 */
cursor?: number | null;
/**
 * @nullable
 */
limit?: number | null;
};